refactor(chat): extract attachment delivery
